+++ eds-user-20210308/alibabacloud_eds_user20210308/models/_create_resource_group_response_body.py
@@ -12,7 +12,7 @@ class CreateResourceGroupResponseBody(DaraModel):
     ):
         # The request ID.
         self.request_id = request_id
-        # The ID of the newly created resource group.
+        # The resource group ID.
         self.resource_group_id = resource_group_id
 
     def validate(self):